vision-microservice/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pycryptodome==3.20.0
ultralytics==8.1.24
paddleocr>=2.7.0.3
paddlepaddle>=2.6.0
boto3==1.34.69
opencv-python-headless==4.9.0.80
numpy<2.0.0
pydantic>=2.6.4
python-multipart==0.0.9
